zingo_status::spend_status::SpendStatus
