+++ zingo-status/src/lib.rs
@@ -2,5 +2,6 @@
 //! TODO: Add Crate Discription Here!
 
 #![warn(missing_docs)]
-#[forbid(unsafe_code)]
+#![forbid(unsafe_code)]
 pub mod confirmation_status;
+pub mod spend_status;
